vim: Provide mechanism for persisting preview window
When using pipe-mysql.vim, I would prefer the preview window stay open
so I can reference the results of the query I just wrote.  But in
general, when I go into insert mode, I'd like preview window to close.

In order to attain this, I extract the behavior out to a function,
provided a mechanism to override the behavior through the presence of a
local variable, and then toggled that variable on in the appropriate
mode.