docker: Convert messy aliases to functions
This commit cleans up the docker aliases by converting most of them to
functions which allows us to target specific containers within a
docker-compose application instead of always targeting the whole setup.
In addition, we've created a functions directory to house related
functions that are sourced by functions.zsh.

TO be able to use the aliases inside functions.zsh, I had to modify
.zshrc to source aliases before functions.