docker: Make helper functions smarter (#19)
This updates the helper commands for docker management to make them a
bit smarter. Instead of assuming docker-compose for some commands and
docker for others, we now try to figure out if the container you're
trying to execute against is docker-compose or docker -- and reacts
accordingly. This provides a seamless experience, allowing you to
dkr-logs the proxy or ssl container as well as containers in your
project.  In addition, you can dkr-zsh/bash in the same manner.

I've also added dkr-restart. dkr-reup stops and starts the container,
reloading any changes to docker-compose files. With dkr-restart, we just
issue a restart command which is **much** faster.

Lastly, I've made dkr-stats smarter. Previously, it showed the stats for
all running containers. While this is useful, it's overkill on systems
running many containers. Sometimes you just want to watch one container
-- and now you can.