zsh: Remove homesick alias
At some point homeshick started using the name of the executable to
determine the configuration directory to use. As a result, we may great
different directories based on whether or not you use this alias.

Additionally, I am going to start installing homeshick into the
appropriate directory instead of calling it homesick.